% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{nst2013}
\alias{nst2013}
\title{US Population Estimates (State Totals: Vintage 2013)}
\format{A data frame containing \code{NAME} for state name, and
population estimates and related metrics from 2010 to 2013.}
\source{
The US Census Bureau:
  \url{http://www.census.gov/popest/data/state/totals/2013}
}
\usage{
nst2013
}
\description{
Population, population change and estimated components of population change:
April 1, 2010 to July 1, 2013 (NST-EST2013-alldata)
}
\examples{
library(shinyCartogram)
str(nst2013)
}
\keyword{datasets}

