% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cartogramOutput}
\alias{cartogramOutput}
\title{Create a \code{svg} element for a Cartogram}
\usage{
cartogramOutput(outputId, width = "100\%", height = "500px",
  topojson = "cartogram/data/us-states-segmentized.topojson",
  colors = rev(brewer.pal(3, "RdYlBu")))
}
\arguments{
\item{outputId}{the id of the \samp{<svg>} element}

\item{width,height}{The width and height of the map. They can either take a
CSS length (e.g. \code{400px} or \code{50\%}) or a numeric value which will
be interpreted as pixels.}

\item{topojson}{The URL for the topojson which encodes the map topology such
as boundaries and names. (the us states are used by default). See
\url{https://github.com/mbostock/topojson/wiki} for information about
where to find other maps or creating your own map.}

\item{colors}{The}
}
\value{
An HTML tag list.
}
\description{
This function is called from \file{ui.R} (or from
\code{\link[shiny]{renderUI}()}); it creates a \code{<svg>} that will contain
a cartogram map.
}

